+++ readium/navigator/src/main/java/org/readium/r2/navigator/VisualNavigator.kt
@@ -14,6 +14,7 @@ import org.readium.r2.navigator.input.InputListener
 import org.readium.r2.navigator.preferences.Axis
 import org.readium.r2.navigator.preferences.ReadingProgression
 import org.readium.r2.shared.ExperimentalReadiumApi
+import org.readium.r2.shared.publication.Link
 import org.readium.r2.shared.publication.Locator
 import org.readium.r2.shared.publication.ReadingProgression as PublicationReadingProgression
 
@@ -81,6 +82,22 @@ public interface VisualNavigator : Navigator {
     public fun removeInputListener(listener: InputListener)
 
     public interface Listener : Navigator.Listener {
+
+        /**
+         * Called when a link to an internal resource was clicked in the navigator.
+         *
+         * You can use this callback to perform custom navigation like opening a new window
+         * or other operations.
+         *
+         * By returning false the navigator wont try to open the link itself and it is up
+         * to the calling app to decide how to display the link.
+         */
+        @Deprecated(
+            "Use `HyperlinkNavigator.Listener.shouldFollowInternalLink` instead",
+            replaceWith = ReplaceWith("shouldFollowInternalLink(link)"),
+            level = DeprecationLevel.ERROR
+        )
+        public fun shouldJumpToLink(link: Link): Boolean { return true }
 
         @Deprecated("Use `addInputListener` instead", level = DeprecationLevel.ERROR)
         public fun onTap(point: PointF): Boolean = false